prefix package with contrib
