+++ contribs/simulatedannealing/src/main/java/org/matsim/contrib/simulatedannealing/SimulatedAnnealingAnalysis.java
@@ -6,7 +6,7 @@
  * the Free Software Foundation; either version 2 of the License,
  * or (at your option) any later version.
  */
-package org.matsim.simulatedannealing;
+package org.matsim.contrib.simulatedannealing;
 
 import org.matsim.core.config.Config;
 import org.matsim.core.controler.MatsimServices;
