+++ contribs/simulatedannealing/src/main/java/org/matsim/contrib/simulatedannealing/cost/CostCalculator.java
@@ -6,17 +6,18 @@
  * the Free Software Foundation; either version 2 of the License,
  * or (at your option) any later version.
  */
-package org.matsim.simulatedannealing;
+package org.matsim.contrib.simulatedannealing.cost;
 
 /**
+ *
+ * Calculates the cost/energy of the (current) solution. Should implement the cost function
+ * that shall be minimised.
+ *
+ * As this is highly problem specific, there is no default implementation.
+ *
  * @author nkuehnel / MOIA
  */
-public interface SimulatedAnnealingListener<T> {
-
-	void solutionAccepted(SimulatedAnnealing.Solution<T> oldSolution, SimulatedAnnealing.Solution<T> newSolution);
-
-	void reset(int iteration);
-
+public interface CostCalculator<T> {
+	double calculateCost(T solution);
 
 }
-
