+++ contribs/simulatedannealing/src/main/java/org/matsim/contrib/simulatedannealing/perturbation/ChainedPeturbatorFactory.java
@@ -6,11 +6,11 @@
  * the Free Software Foundation; either version 2 of the License,
  * or (at your option) any later version.
  */
-package org.matsim.simulatedannealing.perturbation;
+package org.matsim.contrib.simulatedannealing.perturbation;
 
 import org.matsim.contrib.common.util.random.WeightedRandomSelection;
+import org.matsim.contrib.simulatedannealing.SimulatedAnnealing;
 import org.matsim.core.gbl.MatsimRandom;
-import org.matsim.simulatedannealing.SimulatedAnnealing;
 
 import java.util.*;
