+++ contribs/simulatedannealing/src/main/java/org/matsim/contrib/simulatedannealing/perturbation/Perturbator.java
@@ -6,9 +6,9 @@
  * the Free Software Foundation; either version 2 of the License,
  * or (at your option) any later version.
  */
-package org.matsim.simulatedannealing.perturbation;
+package org.matsim.contrib.simulatedannealing.perturbation;
 
-import org.matsim.simulatedannealing.SimulatedAnnealing;
+import org.matsim.contrib.simulatedannealing.SimulatedAnnealing;
 
 /**
  *
