+++ contribs/simulatedannealing/src/main/java/org/matsim/contrib/simulatedannealing/perturbation/PerturbatorFactory.java
@@ -6,10 +6,10 @@
  * the Free Software Foundation; either version 2 of the License,
  * or (at your option) any later version.
  */
-package org.matsim.simulatedannealing.perturbation;
+package org.matsim.contrib.simulatedannealing.perturbation;
 
-import org.matsim.simulatedannealing.SimulatedAnnealing;
-import org.matsim.simulatedannealing.SimulatedAnnealingListener;
+import org.matsim.contrib.simulatedannealing.SimulatedAnnealing;
+import org.matsim.contrib.simulatedannealing.SimulatedAnnealingListener;
 
 /**
  *
