+++ contribs/simulatedannealing/src/main/java/org/matsim/contrib/simulatedannealing/temperature/TemperatureFunction.java
@@ -6,7 +6,7 @@
  * the Free Software Foundation; either version 2 of the License,
  * or (at your option) any later version.
  */
-package org.matsim.simulatedannealing.temperature;
+package org.matsim.contrib.simulatedannealing.temperature;
 
 /**
  *
