+++ contribs/simulatedannealing/src/test/java/org/matsim/contrib/simulatedannealing/DefaultAnnealingAcceptorTest.java
@@ -6,16 +6,14 @@
  * the Free Software Foundation; either version 2 of the License,
  * or (at your option) any later version.
  */
-package org.matsim.simulatedannealing;
+package org.matsim.contrib.simulatedannealing;
 
 import org.apache.commons.lang3.mutable.MutableInt;
 import org.junit.Assert;
 import org.junit.Test;
 import org.matsim.core.controler.IterationCounter;
-import org.matsim.simulatedannealing.SimulatedAnnealing;
-import org.matsim.simulatedannealing.SimulatedAnnealingConfigGroup;
-import org.matsim.simulatedannealing.acceptor.DefaultAnnealingAcceptor;
-import org.matsim.simulatedannealing.temperature.TemperatureFunction;
+import org.matsim.contrib.simulatedannealing.acceptor.DefaultAnnealingAcceptor;
+import org.matsim.contrib.simulatedannealing.temperature.TemperatureFunction;
 
 /**
  * @author nkuehnel / MOIA
